feat: auto parse room directly in admin module
This implements the ability for users to specify either:
  !admin <module> <room> <cmd>
  !admin <module> <cmd>

The former will first resolve <room> to a room id, and pass that to the
module's `admin` function. If <room> cannot be resolved to a valid id,
it will be treated as part of the <cmd> as seen in the latter.

For commands which are not associated with a given room, the new `room`
parameter can be ignored.

The main benefit to this approach is that module authors don't need to
handle their own room parsing. Nor do they need to care whether room was
manually specified or implicitly used based on where the command was
run. They can simply trust that the room that was passed in is the one
that should be acted upon.