Handle room resolving in the admin module (#16)
* feat: pass current room id into the admin function

This is can be used to apply admin actions to the current room, rather
than needing to specify and parse it via the command.

* feat: auto parse room directly in admin module

This implements the ability for users to specify either:
  !admin <module> <room> <cmd>
  !admin <module> <cmd>

The former will first resolve <room> to a room id, and pass that to the
module's `admin` function. If <room> cannot be resolved to a valid id,
it will be treated as part of the <cmd> as seen in the latter.

For commands which are not associated with a given room, the new `room`
parameter can be ignored.

The main benefit to this approach is that module authors don't need to
handle their own room parsing. Nor do they need to care whether room was
manually specified or implicitly used based on where the command was
run. They can simply trust that the room that was passed in is the one
that should be acted upon.

* rename util.rs to room_resolver.rs

* fold duplicate match arms

* update mastodon admin help

* fix bug in mastodon list-posters command

* make room-resolver accept room ids too

Co-authored-by: Benjamin Bouvier <public@benj.me>