refactor: have one Store per Module
The documentation of stores indicates that their lifetime is tied to all
their instances, and that they should be long-lived only if there's a
"main" instance running. Instead, we can have many lightweight instances
with their own store, since they don't interact with each other directly
anyways.

This paves the way for compiling instances in parallel.